+++ core/src/main/kotlin/io/bkbn/kompendium/core/plugin/NotarizedApplication.kt
@@ -1,6 +1,7 @@
 package io.bkbn.kompendium.core.plugin
 
 import io.bkbn.kompendium.core.attribute.KompendiumAttributes
+import io.bkbn.kompendium.json.schema.SchemaConfigurator
 import io.bkbn.kompendium.json.schema.definition.JsonSchema
 import io.bkbn.kompendium.json.schema.util.Helpers.getSimpleSlug
 import io.bkbn.kompendium.oas.OpenApiSpec
@@ -13,7 +14,6 @@ import io.ktor.server.routing.application
 import io.ktor.server.routing.get
 import io.ktor.server.routing.route
 import io.ktor.server.routing.routing
-import kotlin.reflect.KClass
 import kotlin.reflect.KType
 
 object NotarizedApplication {
@@ -28,6 +28,7 @@ object NotarizedApplication {
       }
     }
     var customTypes: Map<KType, JsonSchema> = emptyMap()
+    var schemaConfigurator: SchemaConfigurator = SchemaConfigurator.Default()
   }
 
   operator fun invoke() = createApplicationPlugin(
@@ -41,5 +42,6 @@ object NotarizedApplication {
       spec.components.schemas[type.getSimpleSlug()] = schema
     }
     application.attributes.put(KompendiumAttributes.openApiSpec, spec)
+    application.attributes.put(KompendiumAttributes.schemaConfigurator, pluginConfig.schemaConfigurator)
   }
 }