+++ json-schema/src/main/kotlin/io/bkbn/kompendium/json/schema/handler/MapHandler.kt
@@ -1,6 +1,7 @@
 package io.bkbn.kompendium.json.schema.handler
 
 import io.bkbn.kompendium.json.schema.SchemaGenerator
+import io.bkbn.kompendium.json.schema.SchemaConfigurator
 import io.bkbn.kompendium.json.schema.definition.JsonSchema
 import io.bkbn.kompendium.json.schema.definition.MapDefinition
 import io.bkbn.kompendium.json.schema.definition.NullableDefinition
@@ -14,12 +15,12 @@ import kotlin.reflect.KType
 
 object MapHandler {
 
-  fun handle(type: KType, cache: MutableMap<String, JsonSchema>): JsonSchema {
+  fun handle(type: KType, cache: MutableMap<String, JsonSchema>, schemaConfigurator: SchemaConfigurator): JsonSchema {
     require(type.arguments.first().type?.classifier as KClass<*> == String::class) {
       "JSON requires that map keys MUST be Strings.  You provided ${type.arguments.first().type}"
     }
     val valueType = type.arguments[1].type ?: error("this indicates a bug in Kompendium, please open a GitHub issue")
-    val valueSchema = SchemaGenerator.fromTypeToSchema(valueType, cache).let {
+    val valueSchema = SchemaGenerator.fromTypeToSchema(valueType, cache, schemaConfigurator).let {
       if (it is TypeDefinition && it.type == "object") {
         cache[valueType.getSimpleSlug()] = it
         ReferenceDefinition(valueType.getReferenceSlug())