+++ json-schema/src/main/kotlin/io/bkbn/kompendium/json/schema/handler/SealedObjectHandler.kt
@@ -1,6 +1,7 @@
 package io.bkbn.kompendium.json.schema.handler
 
 import io.bkbn.kompendium.json.schema.SchemaGenerator
+import io.bkbn.kompendium.json.schema.SchemaConfigurator
 import io.bkbn.kompendium.json.schema.definition.AnyOfDefinition
 import io.bkbn.kompendium.json.schema.definition.JsonSchema
 import io.bkbn.kompendium.json.schema.definition.ReferenceDefinition
@@ -13,11 +14,16 @@ import kotlin.reflect.full.createType
 
 object SealedObjectHandler {
 
-  fun handle(type: KType, clazz: KClass<*>, cache: MutableMap<String, JsonSchema>): JsonSchema {
+  fun handle(
+    type: KType,
+    clazz: KClass<*>,
+    cache: MutableMap<String, JsonSchema>,
+    schemaConfigurator: SchemaConfigurator
+  ): JsonSchema {
     val subclasses = clazz.sealedSubclasses
       .map { it.createType(type.arguments) }
       .map { t ->
-        SchemaGenerator.fromTypeToSchema(t, cache).let { js ->
+        SchemaGenerator.fromTypeToSchema(t, cache, schemaConfigurator).let { js ->
           if (js is TypeDefinition && js.type == "object") {
             cache[t.getSimpleSlug()] = js
             ReferenceDefinition(t.getReferenceSlug())