+++ json-schema/src/main/kotlin/io/bkbn/kompendium/json/schema/handler/SimpleObjectHandler.kt
@@ -1,6 +1,7 @@
 package io.bkbn.kompendium.json.schema.handler
 
 import io.bkbn.kompendium.json.schema.SchemaGenerator
+import io.bkbn.kompendium.json.schema.SchemaConfigurator
 import io.bkbn.kompendium.json.schema.definition.JsonSchema
 import io.bkbn.kompendium.json.schema.definition.NullableDefinition
 import io.bkbn.kompendium.json.schema.definition.OneOfDefinition
@@ -14,22 +15,29 @@ import kotlin.reflect.KType
 import kotlin.reflect.KTypeParameter
 import kotlin.reflect.KTypeProjection
 import kotlin.reflect.full.createType
-import kotlin.reflect.full.memberProperties
 import kotlin.reflect.full.primaryConstructor
+import kotlin.reflect.jvm.javaField
 
 object SimpleObjectHandler {
 
-  fun handle(type: KType, clazz: KClass<*>, cache: MutableMap<String, JsonSchema>): JsonSchema {
+  fun handle(
+    type: KType,
+    clazz: KClass<*>,
+    cache: MutableMap<String, JsonSchema>,
+    schemaConfigurator: SchemaConfigurator
+  ): JsonSchema {
 
     cache[type.getSimpleSlug()] = ReferenceDefinition(type.getReferenceSlug())
 
     val typeMap = clazz.typeParameters.zip(type.arguments).toMap()
-    val props = clazz.memberProperties.associate { prop ->
+    val props = schemaConfigurator.serializableMemberProperties(clazz)
+      .filterNot { it.javaField == null }
+      .associate { prop ->
       val schema = when (prop.needsToInjectGenerics(typeMap)) {
-        true -> handleNestedGenerics(typeMap, prop, cache)
+        true -> handleNestedGenerics(typeMap, prop, cache, schemaConfigurator)
         false -> when (typeMap.containsKey(prop.returnType.classifier)) {
-          true -> handleGenericProperty(prop, typeMap, cache)
-          false -> handleProperty(prop, cache)
+          true -> handleGenericProperty(prop, typeMap, cache, schemaConfigurator)
+          false -> handleProperty(prop, cache, schemaConfigurator)
         }
       }
 
@@ -38,14 +46,16 @@ object SimpleObjectHandler {
         false -> schema
       }
 
-      prop.name to nullCheckSchema
+      schemaConfigurator.serializableName(prop) to nullCheckSchema
     }
 
-    val required = clazz.memberProperties.filterNot { prop -> prop.returnType.isMarkedNullable }
+    val required = schemaConfigurator.serializableMemberProperties(clazz)
+      .asSequence()
+      .filterNot { it.javaField == null }
+      .filterNot { prop -> prop.returnType.isMarkedNullable }
       .filterNot { prop -> clazz.primaryConstructor!!.parameters.find { it.name == prop.name }!!.isOptional }
-      .map { it.name }
+      .map { schemaConfigurator.serializableName(it) }
       .toSet()
-
 
     val definition = TypeDefinition(
       type = "object",
@@ -69,7 +79,8 @@ object SimpleObjectHandler {
   private fun handleNestedGenerics(
     typeMap: Map<KTypeParameter, KTypeProjection>,
     prop: KProperty<*>,
-    cache: MutableMap<String, JsonSchema>
+    cache: MutableMap<String, JsonSchema>,
+    schemaConfigurator: SchemaConfigurator
   ): JsonSchema {
     val propClass = prop.returnType.classifier as KClass<*>
     val types = prop.returnType.arguments.map {
@@ -77,7 +88,7 @@ object SimpleObjectHandler {
       typeMap.filterKeys { k -> k.name == typeSymbol }.values.first()
     }
     val constructedType = propClass.createType(types)
-    return SchemaGenerator.fromTypeToSchema(constructedType, cache).let {
+    return SchemaGenerator.fromTypeToSchema(constructedType, cache, schemaConfigurator).let {
       if (it.isOrContainsObjectDef()) {
         cache[constructedType.getSimpleSlug()] = it
         ReferenceDefinition(prop.returnType.getReferenceSlug())
@@ -90,11 +101,12 @@ object SimpleObjectHandler {
   private fun handleGenericProperty(
     prop: KProperty<*>,
     typeMap: Map<KTypeParameter, KTypeProjection>,
-    cache: MutableMap<String, JsonSchema>
+    cache: MutableMap<String, JsonSchema>,
+    schemaConfigurator: SchemaConfigurator
   ): JsonSchema {
     val type = typeMap[prop.returnType.classifier]?.type
       ?: error("This indicates a bug in Kompendium, please open a GitHub issue")
-    return SchemaGenerator.fromTypeToSchema(type, cache).let {
+    return SchemaGenerator.fromTypeToSchema(type, cache, schemaConfigurator).let {
       if (it.isOrContainsObjectDef()) {
         cache[type.getSimpleSlug()] = it
         ReferenceDefinition(type.getReferenceSlug())
@@ -104,8 +116,12 @@ object SimpleObjectHandler {
     }
   }
 
-  private fun handleProperty(prop: KProperty<*>, cache: MutableMap<String, JsonSchema>): JsonSchema =
-    SchemaGenerator.fromTypeToSchema(prop.returnType, cache).let {
+  private fun handleProperty(
+    prop: KProperty<*>,
+    cache: MutableMap<String, JsonSchema>,
+    schemaConfigurator: SchemaConfigurator
+  ): JsonSchema =
+    SchemaGenerator.fromTypeToSchema(prop.returnType, cache, schemaConfigurator).let {
       if (it.isOrContainsObjectDef()) {
         cache[prop.returnType.getSimpleSlug()] = it
         ReferenceDefinition(prop.returnType.getReferenceSlug())