+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+- Support for @Transient annotation
+- Support for @SerialName annotation on fields
+- Supports for un-backed fields, by excluding them from the generated schema.
+
 ### Added
 
 ### Changed